Rename FeaturedCard and split styles out
